fix(build): moved speaker list to page
